refactor: table-logic from file to module
- first draft of new factory class
- orders components like columns and util functions into files in module
